Clear avatar media cache on theme change
Default avatars retained the previou;sly applied theme's colors
due to the avatar images being cached.  This patch checks if a
theme change has occurred and if so destroys the avatar media cache
so that it can be rebuilt with the newly applied theme's colors.

Change-Id: Iab3d6e0171a949235da643fc80bf54d9375e42fb
TICKET: CYNGNOS-3118

Conflicts:
    src/com/android/messaging/BugleApplication.java

DarkKat changes:
This patch was originally made for theme changes via CyanogenMod/T-Mobile Theme Engine,
so for theming via day/night themes,
we have to check for "CONFIG_UI_MODE" instead of "CONFIG_THEME_RESOURCE".